+++ src/Decoder.ts
@@ -1,7 +1,7 @@
 import { prettyByte } from "./utils/prettyByte";
 import { ExtensionCodec } from "./ExtensionCodec";
 import { getInt64, getUint64 } from "./utils/int";
-import { utf8DecodeJs, TEXT_DECODER_AVAILABLE, TEXT_DECODER_THRESHOLD, utf8DecodeTD } from "./utils/utf8";
+import { utf8DecodeJs, TEXT_ENCODING_AVAILABLE, TEXT_DECODER_THRESHOLD, utf8DecodeTD } from "./utils/utf8";
 import { createDataView, ensureUint8Array } from "./utils/typedArrays";
 import { WASM_AVAILABLE, WASM_STR_THRESHOLD, utf8DecodeWasm } from "./wasmFunctions";
 
@@ -68,7 +68,7 @@ export class Decoder {
     readonly maxExtLength = DEFAULT_MAX_LENGTH,
   ) {}
 
-  setBuffer(buffer: ArrayLike<number>): void {
+  setBuffer(buffer: ArrayLike<number> | ArrayBuffer): void {
     this.bytes = ensureUint8Array(buffer);
     this.view = createDataView(this.bytes);
     this.pos = 0;
@@ -482,7 +482,7 @@ export class Decoder {
 
     const offset = this.pos + headerOffset;
     let object: string;
-    if (TEXT_DECODER_AVAILABLE && byteLength > TEXT_DECODER_THRESHOLD) {
+    if (TEXT_ENCODING_AVAILABLE && byteLength > TEXT_DECODER_THRESHOLD) {
       object = utf8DecodeTD(this.bytes, offset, byteLength);
     } else if (WASM_AVAILABLE && byteLength > WASM_STR_THRESHOLD) {
       object = utf8DecodeWasm(this.bytes, offset, byteLength);
@@ -502,7 +502,8 @@ export class Decoder {
       throw MORE_DATA;
     }
 
-    const object = new Uint8Array(this.view.buffer, this.view.byteOffset + this.pos + headOffset, byteLength);
+    const offset = this.pos + headOffset;
+    const object = this.bytes.subarray(offset, offset + byteLength);
     this.pos += headOffset + byteLength;
     return object;
   }