+++ src/timestamp.ts
@@ -8,13 +8,13 @@ export type TimeSpec = {
   nsec: number;
 };
 
-const TIMESTAMP32_MAX_SEC = 0x100000000; // 32-bit unsigned int
-const TIMESTAMP64_MAX_SEC = 0x400000000; // 34-bit unsigned int
+const TIMESTAMP32_MAX_SEC = 0x100000000 - 1; // 32-bit unsigned int
+const TIMESTAMP64_MAX_SEC = 0x400000000 - 1; // 34-bit unsigned int
 
 export function encodeTimeSpecToTimestamp({ sec, nsec }: TimeSpec): Uint8Array {
-  if (sec >= 0 && nsec >= 0 && sec < TIMESTAMP64_MAX_SEC) {
+  if (sec >= 0 && nsec >= 0 && sec <= TIMESTAMP64_MAX_SEC) {
     // Here sec >= 0 && nsec >= 0
-    if (nsec === 0 && sec < TIMESTAMP32_MAX_SEC) {
+    if (nsec === 0 && sec <= TIMESTAMP32_MAX_SEC) {
       // timestamp 32 = { sec32 (unsigned) }
       const rv = new Uint8Array(4);
       const view = new DataView(rv.buffer);
@@ -65,19 +65,18 @@ export function encodeTimestampExtension(object: unknown): Uint8Array | null {
 }
 
 export function decodeTimestampToTimeSpec(data: Uint8Array): TimeSpec {
+  const view = new DataView(data.buffer, data.byteOffset, data.byteLength);
+
   // data may be 32, 64, or 96 bits
   switch (data.byteLength) {
     case 4: {
       // timestamp 32 = { sec32 }
-      const view = new DataView(data.buffer, data.byteOffset, data.byteLength);
       const sec = view.getUint32(0);
       const nsec = 0;
       return { sec, nsec };
     }
     case 8: {
       // timestamp 64 = { nsec30, sec34 }
-      const view = new DataView(data.buffer, data.byteOffset, data.byteLength);
-
       const nsec30AndSecHigh2 = view.getUint32(0);
       const secLow32 = view.getUint32(4);
       const sec = (nsec30AndSecHigh2 & 0x3) * 0x100000000 + secLow32;
@@ -86,7 +85,6 @@ export function decodeTimestampToTimeSpec(data: Uint8Array): TimeSpec {
     }
     case 12: {
       // timestamp 96 = { nsec32 (unsigned), sec64 (signed) }
-      const view = new DataView(data.buffer, data.byteOffset, data.byteLength);
 
       const sec = getInt64(view, 4);
       const nsec = view.getUint32(0);